Move wasm crypto tests to client
